various fixes to avoid undefined behavior for binary32 functions
(contributed by Ganesh Ajjanagadde <gajjanag@alum.mit.edu>
 see https://sympa.inria.fr/sympa/arc/core-math/2025-03/msg00007.html)

+++ src/binary32/cospi/cospif.c
@@ -93,10 +93,10 @@ float cr_cospif(float x){
   }
   if(__builtin_expect(p>31, 0)) {
     if(__builtin_expect(p>63, 0)) return 1.0f;
-    int32_t iq = m << (p - 32);
+    int32_t iq = (uint32_t)m << (p - 32);
     return S[(iq+32)&127];
   }
-  int32_t k = m << p;
+  int32_t k = (uint32_t)m << p;
   if(__builtin_expect(k==0, 0)){
     int32_t iq = m >> (32-p);
     return S[(iq+32)&127];
